saas-data-migration/requirements.txt
azure-core==1.30.0
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
isodate==0.6.1
numpy==1.26.4
pandas==2.2.1
pycparser==2.21
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
